다중 파일 관련 클래스 및 유틸 파일들 까지 작업

+++ src/main/java/com/abc/domain/FreeBoard.java
@@ -15,7 +15,7 @@ import lombok.ToString;
 @ToString
 public class FreeBoard {
 
-	private int boardNum;
+	private long boardNum;
 	private String memberNum;
 	private String nickname;
 	private String category;
